fix lints and remove by rony requests
