+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py37,py38,flake8,mypy
+envlist = py37,py38,flake8
 
 [testenv]
 passenv = *
@@ -11,17 +11,7 @@ commands = pytest -p no:warnings -v {posargs}
 basepython = python3
 skip_install = true
 deps = flake8
-commands = flake8 demisto_sdk/ tests/ setup.py
-
-[testenv:mypy]
-basepython = python3
-ignore_outcome = true
-skip_install = true
-deps = mypy
-commands =
-    mypy --check-untyped-defs --ignore-missing-imports \
-    --follow-imports=silent --show-column-numbers --show-error-codes --pretty \
-    --allow-redefinition demisto_sdk/ setup.py
+commands = flake8 demisto_sdk/ setup.py
 
 [flake8]
 ignore = D203, W293, W504, W291, W605, F405, F403
